displayConfirmationModal is not a component -> move it to utils/admin
