+++ assembl/static2/js/app/utils/administration/displayConfirmationModal.jsx
@@ -2,8 +2,8 @@
 import React from 'react';
 import { Translate } from 'react-redux-i18n';
 import { Button } from 'react-bootstrap';
-import { displayModal, closeModal } from '../../utils/utilityManager';
-import type { TInitialValues } from '../form/LoadSaveReinitializeForm';
+import { displayModal, closeModal } from '../utilityManager';
+import type { TInitialValues } from '../../components/form/LoadSaveReinitializeForm';
 
 // This is a generic function to display a modal in
 // an administration form when a confirmation of the changes is required
